Move `CommentCellController` to its own file
