+++ HackrNewsiOS/Story Details UI/Controllers/StoryDetailsViewController.swift
@@ -5,37 +5,6 @@
 import HackrNews
 import UIKit
 
-public protocol CommentCellControllerDelegate {
-    func didRequestComment()
-}
-
-public class CommentCellController: CommentView, CommentLoadingView, CommentErrorView {
-    private(set) var cell: CommentCell?
-    private let delegate: CommentCellControllerDelegate
-
-    public init(delegate: CommentCellControllerDelegate) {
-        self.delegate = delegate
-    }
-
-    func view(in tableView: UITableView) -> CommentCell {
-        cell = tableView.dequeueReusableCell()
-        delegate.didRequestComment()
-        return cell!
-    }
-
-    public func display(_ viewModel: CommentViewModel) {
-        cell?.authorLabel.text = viewModel.author
-        cell?.createdAtLabel.text = viewModel.createdAt
-        cell?.bodyLabel.text = viewModel.text
-    }
-
-    public func display(_ viewModel: CommentLoadingViewModel) {
-        cell?.isLoadingContent = viewModel.isLoading
-    }
-
-    public func display(_: CommentErrorViewModel) {}
-}
-
 public class StoryDetailsViewController: UITableViewController {
     private var detailsSection: Int { 0 }
     private var storyCell: IndexPath { IndexPath(row: 0, section: detailsSection) }
